refactor(webclient): move opening-hours editor into the properties modal
